+++ NEWS.md
@@ -3,11 +3,14 @@ The current version of the `MSEtool` package is available for download from [CRA
 # MSEtool 3.4.9999 - Development Version
 
 ## New additions
+- argument `parallel` in `runMSE` can takes a named list of MPs
 - two new functions `WHAM2OM` and `ASAP2OM` to import assessments into operating models
+- `Assess2MOM` to generate a multi-fleet operating model, either for a 2-sex population where recruitment is predicted from female spawning output, or a single-sex population.
 
 ## Minor changes
 - add message showing range of estimated von Bert. parameters when `OM@cpars$Len_age` is provided.
 - fix bug in SSBpR for 2-sex `MOM` (introduced in 3.4.0)
+- pass projected OM abundance and SSB arrays to `Data@Misc$StockPars` (slot `N_P`, `SSB_P`, etc.) during closed-loop
 
 # MSEtool 3.4.0